Use greedy matching to handle case where lines of text are repeated in transcript.
Previously LLM output was stored as a dict where each line was assigned one text_type. This was problematic when a line of text was repeated on the page because (1) the text_type could be different in different occurences, and (2) once a line matched a text_type in the dict, the entire k,v pair was removed, which meant any future repetitions of the line had no text_type match. Now we store text_types as a list in order of occurrence associated with the line, and as we go through the gold standard transcript, we compare the true text_type to the next LLM-generated text_type in the list.